Update cgreen version hash
Prior to this commit the version of Cgreen checked out in the
`.travis.yml` file introduced a memory leak to the CgreenDriver.
After this commit the version of Cgreen checked out by Travis reflects
a version of Cgreen with that memory leak fixed. Due to the bump in
Cgreen version that introduced new cgreen error output,
some "black box" functionality of the cucumber-cpp CgreeDriver was being reported.
This commit also truncates that information so it isn't visible in a failed tests'
error output.