Fortify code, address change requests.
Prior to this commit this feature's implementation did not fully conform
to C++ standards and convention. This lead to inconsistencies and
ambiguity in the source. After this commit the implementation should
conform somewhat more to this project's standards as well as a C++ convention.
Other changes included in this commit:
- cmake revisions to match the functionality and usage for this project's gmock dependency.
- travis file cleanup, the logic applied to Cgreen steps did not fully
  conform to the project's other dependencies.
- Documentation regarding supported Cgreen versions (or more accurately
  lack of versions)